Moved permissions functionality to its own component
https://issues.redhat.com/browse/COST-926
