+++ src/components/async/asyncComponent/asyncComponent.tsx
