+++ src/components/async/asyncComponent/index.ts
@@ -1,0 +1,3 @@
+import { asyncComponent } from './asyncComponent';
+
+export default asyncComponent;
