+++ src/components/async/permissionsComponent/index.ts
@@ -1,0 +1,3 @@
+import { Permissions } from './permissions';
+
+export default Permissions;
